extracted imageupload methods from article into mixin
